Option to include and exclude an index
The default is `index=None`, which only adds an index for
pandas.DataFrames. If `index` is an interable, use that as the
index (`len(index)` must be the same as the number of rows).

If `index=True`, add a running number in front of each row or
the dataframe index, if the input data is a pandas.DataFrame. If
`index=False`, no index is added, even for pandas.DataFrames.