Add `pretty` format to behave like PrettyTable
This adds a new `pretty` default TableFormat to make it easy to convert
legacy code using the PrettyTables library over to using tabulate.

It does not provide 100% output compatibility, but it does address many
of the differences between PrettyTable's output and the closest existing
one of psql. This will help minimize the impact in unit tests and
expected user output for projects to make the switch.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>